fix: Align lines with emojis correctly
Alignment issues occur when some lines have certain emojis in them. This
is caused by incorrectly calculating the width of a string.

For a string with emojis in them, it is important to understand that
the string width is different to the rune width. Many functions within
Lipgloss base the string length on the printable rune width which is
incorrect. The correct method requires finding the string width.

The reflow package has a printable rune width function. While the ANSI
sequences need to be removed, we do not want the rune width as this
function returns.

This change adds a new function to get the printable string width. The
code was taken from reflow and modified for use in Lipgloss. Since the
printable rune width function in reflow works correctly, it was more
reasonable to add the function to Lipgloss.