feat: add adaptive color package
This introduces a helper type `LightDark` that takes a boolean to
determine which `Color(light, dark)` to choose from. The `adaptive`
package is a helper package that uses the `lipgloss.LightDark` along
with querying the terminal when the module is imported to choose the
appropriate light-dark color.

Example:

```go
var (
  light = "#0000ff"
  dark = "#ff0000"
)

colorToUse := adaptive.Color(light, dark) // the terminal is queried before choosing the color
fmt.Println(colorToUse)
```