fix: manually query terminal for background color
This drops the dependency on `github.com/charmbracelet/x/input` and
`github.com/erikgeiser/coninput` by manually parsing the terminal
response to the background color query using `ansi`.

Since we moved `x/input` to Bubble Tea, and the code is no longer
maintained, we need a different approach to query the terminal for the
background color. This is basically doing the same thing as before, but
manually parsing the response instead of using the `x/input` package.

Fixes: https://github.com/charmbracelet/x/issues/248
Supersedes: https://github.com/charmbracelet/x/pull/249